# PERF: GH10213 kth_smallest GIL release