@jreback 2015-08-29
can you add a asv benchmark (note that #10928 ) changes this so will merge this after.

This *should* affect ``median``, ``nlargest``, ``nsmallest``. So those would be nice to test.

Further use the ``test_parallel`` decorator to show the perf changes.