@kawochen 2015-08-29
OK, I will update after 10928.  I had looked at `test_parallel`, but wasn't sure how to make it work for this function. 

The function alters the state of the array, so I needed one set-up per iteration (hence `repeat=n` and `number=1`), but obviously didn't want that to be included in the timing.  In-place sorting and functions that drop stuff should have the same problem, so maybe there's machinery to deal with this that I don't know about.  Plus `test_parallel` has a function level import, which I thought might skew the timing for this one. 

I will check what asv has.