@kawochen 2015-09-01
Will do.  I haven't been able to get the `asv` benchmarks to match the improvement I saw, even when I add class attributes like `number` and `repeat`.  `asv` does use a different default timer, but it looked more like `number` and `repeat` are ignored.  I will update tonight.