@pv 2015-09-03
asv `repeat` does not actually re-run setup each time, which probably should be considered a bug. (If you want to elaborate on the other million gotchas, that could also be useful)