@kawochen 2015-09-03
@pv forgive my hyperbole! It's a great package :smile:   I looked at the code before and saw that repeat is passed into `timeit`, but you are saying it actually doesn't quite work?  Is there an open issue for this?  I'll be more constructive and translate some of what I think are gotchas into feature requests.