@pv 2015-09-03
Repeat is indeed passed to timeit, but the setup function is not run by timeit, but is done separately.
cf https://github.com/spacetelescope/asv/pull/316